+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -4,8 +4,11 @@ Release History
 ===============
 2.2.12
 ++++++
+* 'vm disk/unmanaged-disk' and 'vmss disk': parameter options are more consistent.
 * `vm/vmss create`: support cross tenant image referencing.
 * `vm diagnostics get-default-config --windows-os`: fix bug in windows os default config.
+* `vmss extension set`: expose `--provision-after-extensions`. It defines what extensions must be provisioned before the extension being set.
+* `sig image-version update`: expose --replica-count for setting the default replication count.
 
 2.2.11
 ++++++